LPS-68601 Moves to internal
